DAS-7183: Take all date logic to DateRangeFilter file
